+++ tests/StackExchange.Redis.Tests/OverloadCompat.cs
@@ -15,7 +15,41 @@ public class OverloadCompat : TestBase
     public OverloadCompat(ITestOutputHelper output, SharedConnectionFixture fixture) : base (output, fixture) { }
 
     [Fact]
-    public async Task StringGet()
+    public async Task KeyExpire()
+    {
+        using var conn = Create();
+        var db = conn.GetDatabase();
+        RedisKey key = Me();
+        var expiresIn = TimeSpan.FromSeconds(10);
+        var expireTime = DateTime.UtcNow.AddHours(1);
+        var when = ExpireWhen.Always;
+        var flags = CommandFlags.None;
+
+        db.KeyExpire(key, expiresIn);
+        db.KeyExpire(key, expiresIn, when);
+        db.KeyExpire(key, expiresIn, flags);
+        db.KeyExpire(key, expiresIn, when, flags);
+
+        db.KeyExpire(key, expireTime);
+        db.KeyExpire(key, expireTime, when);
+        db.KeyExpire(key, expireTime, flags);
+        db.KeyExpire(key, expireTime, when, flags);
+
+        // Async
+
+        await db.KeyExpireAsync(key, expiresIn);
+        await db.KeyExpireAsync(key, expiresIn, when);
+        await db.KeyExpireAsync(key, expiresIn, flags);
+        await db.KeyExpireAsync(key, expiresIn, when, flags);
+
+        await db.KeyExpireAsync(key, expireTime);
+        await db.KeyExpireAsync(key, expireTime, when);
+        await db.KeyExpireAsync(key, expireTime, flags);
+        await db.KeyExpireAsync(key, expireTime, when, flags);
+    }
+
+    [Fact]
+    public async Task StringSet()
     {
         using var conn = Create();
         var db = conn.GetDatabase();